Refactor: project structure and modularize code
This commit refactors the project structure and modularizes the codebase for better organization and maintainability. The main changes include:

Separating the project into multiple modules:
main.py remains the entry point of the application, but its functionality has been simplified to initialize the CLI and start the setup process using the SetupProject class.
setup_project.py defines the setup process for creating the boilerplate. It takes user inputs for project name, framework choice, and setup choice.
create_project.py contains the logic to create the project directory and copy the template files based on the user's choices. It also provides usage instructions.
delete_project.py handles the option to delete/override an existing project directory if it already exists.
constants.py defines constants such as the template folder and current directory, improving code readability.
