+++ myte/constants.py
@@ -1,0 +1,10 @@
+# constants.py
+
+"""
+This module defines constants use for this project
+"""
+
+import os
+
+template_folder = os.path.join(os.path.dirname(__file__), "templates")
+current_dir = os.getcwd()
